[SB-31] Webapp > Sidebar fix
