+++ src/components/Sidebar/index.js
@@ -1,3 +1,3 @@
-import Sidebar from './Sidebar';
+import Sidebar from './Sidebar.container';
 
 export default Sidebar;
